+++ rust-toolchain
@@ -1,1 +1,1 @@
-nightly-2021-06-06
+nightly-2021-10-28
