+++ rust/Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 members = [
-    "bls12_381-sign", "grpc-server"
+    "bls12_381-sign", "bls12_381-sign-ffi", "grpc-server"
 ]
 
 [profile.dev]
