+++ rust/bls12_381-sign/Cargo.toml
@@ -1,29 +1,21 @@
 [package]
 name = "dusk-bls12_381-sign"
-version = "0.1.0-rc.0"
+version = "0.1.0-rc.1"
 authors = [
     "Jules de Smit <julesdesmit@gmail.com>",
     "Matteo Ferretti <matteo@dusk.network>",
     "Emanuele Paterna <seppia@dusk.network>",
     "David Vennik <david@dusk.network>",
 ]
-edition = "2018"
+edition = "2021"
 description = "Implementation of BLS signatures using the BLS12-381 curve"
 license = "MPL-2.0"
 
 [dependencies]
-dusk-bls12_381 = { version = "0.8", default-features = false, features = ["groups"] }
+blake2 = { version = "0.10", default-features = false }
+dusk-bls12_381 = { version = "0.8", default-features = false, features = ["pairings"] }
 dusk-bytes = "0.1"
-rand_core = {version = "0.6"}
-blake2 = {version = "0.9", default-features = false, optional = true}
-libc = "0.2"
+rand_core = { version = "0.6", default-features = false }
 
-[lib]
-crate-type = ["rlib", "staticlib"]
-
-[features]
-default = ["std"]
-std = [
-    "dusk-bls12_381/default",
-    "blake2",
-]
+[dev-dependencies]
+rand = { version = "0.8", default-features = false, features = ["std_rng"] }
