+++ rust/bls12_381-sign/src/error.rs
@@ -4,8 +4,9 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-use core::fmt;
 use dusk_bytes::Error as DuskBytesError;
+
+use core::fmt;
 
 /// Standard error for the interface
 #[derive(Copy, Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
