+++ rust/rustfmt.toml
@@ -1,3 +1,2 @@
 max_width = 80
 wrap_comments = true
-newline_style = "unix"
